LayoutTests/fast/history/page-cache-createImageBitmap-expected.txt

Tests that a page with a pending createImageBitmap promise is able to enter the page cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS - Page did enter and was restored from the page cache
PASS !!imageBitmap is true
PASS restoredFromPageCache is true
PASS successfullyParsed is true

TEST COMPLETE
